services/user-service/requirements.txt
fastapi
uvicorn[standard]
pydantic>=2
SQLAlchemy>=2
psycopg2-binary
python-jose[cryptography]